try 1

+++ modules/enkf/perturbation/src/mod_random_forcing.F90
@@ -121,8 +121,8 @@ contains
       implicit none
       integer:: perturbation_count
       real :: dx
-      real*8, dimension( 2,xdim*ydim) :: synforc00, synforc01
-      real*8, dimension(10,xdim*ydim) :: randfld00, randfld01
+      real*4, dimension( 2,xdim*ydim) :: synforc00, synforc01
+      real*4, dimension(10,xdim*ydim) :: randfld00, randfld01
       if(.not.randf) then
         write(*,'("randf option switched off in pseudo2D.nml,no perturbation will be applied")')
         return
@@ -309,7 +309,7 @@ contains
       real, parameter :: wlat=60.
       integer i,j
       real*8, save :: rdtime=8.d0/24.d0    ! Time step of forcing update
-      real*8  :: randfld(idm*jdm,10), synforc(idm*jdm,2)
+      real*4  :: randfld(idm*jdm,10), synforc(idm*jdm,2)
       ! Autocorrelation between two times "tcorr"
       !KAL - quite high? - autocorr = 0.95
       autocorr = exp(-1.0)
@@ -608,7 +608,7 @@ contains
 !------------------------------------------
    subroutine load_randfld_synforc(randfld, synforc) ! todo, check if it is necessary to read synforc
       integer :: ix,jy,id
-      real*8  :: randfld(idm*jdm, 10), synforc(idm*jdm,2)
+      real*4  :: randfld(idm*jdm, 10), synforc(idm*jdm,2)
       do jy=1,jdm
       do ix=1,idm
          id = (jy-1)*idm + ix
@@ -634,7 +634,7 @@ contains
    !------------------------------------------
    subroutine save_randfld_synforc(randfld, synforc)
       integer :: ix,jy,id
-      real*8  :: randfld(idm*jdm,10), synforc(idm*jdm,2)
+      real*4  :: randfld(idm*jdm,10), synforc(idm*jdm,2)
       do jy=1,jdm
       do ix=1,idm
          id = (jy-1)*idm + ix
@@ -799,7 +799,7 @@ contains
          character(2)  :: time_index
          character(150) :: filename
          integer       :: ix,jy,id
-        real*8  :: randfld(idm*jdm,10), synforc(idm*jdm,2)
+        real*4  :: randfld(idm*jdm,10), synforc(idm*jdm,2)
          filename = trim(iopath)//'/synforc.'//time_index
 
          if (debug) print*, 'writing synforc ', filename
